Use xvfb-run for running tests inside Travis
459b9fc068c2a34585455190d75d5254c1be1ed5 removed the 'script' as Travis would use 'npm test' by
default anyways. Initially this worked, but it seems that something has changed and we're now seeing
timeouts, likely because the browsers do not properly start up. For now: switch back to using 'xvfb-run'
to make the tests work.

Note that we do not want this xvfb-run encoded in the package.json 'test' script, because that would
a) break windows compatibility, b) require local changes when debugging test.